[tx] add submitRegisterParticipant
